requirements.txt
cffi==1.17.1
cryptography==43.0.1
pycparser==2.22
pycryptodome==3.20.0
